pyproject.toml
[tool.poetry]
name = "python-autotest-lesson-mobile"
version = "0.1.0"
description = ""
authors = ["Julia Ivanova <murova47@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.3"
Appium-Python-Client = "^2.6.1"
selene = {version = "^2.0.0-beta.10", allow-prereleases = true}
allure-pytest = "^2.10.0"
requests = "^2.28.1"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
pytest-xdist = "^2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts="-vv --alluredir=allure-results --clean-alluredir"
